Backend/requirements.txt
sqlalchemy>=1.4.0
alembic>=1.7.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
pydantic>=1.8.0
fastapi>=0.68.0
uvicorn>=0.15.0
email-validator>=2.0.0
httpx>=0.23.0
google-generativeai>=0.3.0
python-docx>=0.8.11